Update action mySQL
ID of user should be larger.

+++ modules/nvform/action_mysql.php
@@ -44,7 +44,7 @@ $sql_create_module[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_
   fid mediumint(8) NOT NULL DEFAULT '0',
   answer text,
   answer_extend text,
-  who_answer tinyint(2) NOT NULL DEFAULT '0',
+  who_answer int(11) NOT NULL DEFAULT '0',
   answer_time int(11) NOT NULL DEFAULT '0',
   answer_edit_time int(11) NOT NULL DEFAULT '0',
   PRIMARY KEY (id)
